Switch SQS to AWS SDK v2
For consistency and to enable tracing through SQS

Fixes SSM-41 and SSM-84 #patch
